+++ sdk/servicebus/azure-servicebus/azure/servicebus/aio/_transport/_base_async.py
@@ -39,6 +39,8 @@ class AmqpTransportAsync(ABC):  # pylint: disable=too-many-public-methods
     AMQP_LONG_VALUE: Callable
     AMQP_ARRAY_VALUE: Callable
     AMQP_UINT_VALUE: Callable
+    AMQP_INT_VALUE: Callable
+    AMQP_TIMESTAMP_VALUE: Callable
 
     @staticmethod
     @abstractmethod
@@ -308,6 +310,16 @@ class AmqpTransportAsync(ABC):  # pylint: disable=too-many-public-methods
          then pass 300 to refresh_window. Only used by uamqp.
         :returns: JWTTokenAuth.
         :rtype: ~pyamqp.aio._authentication_async.JWTTokenAuth or ~uamqp.authentication.JWTTokenAuth
+        """
+
+    @staticmethod
+    @abstractmethod
+    async def mgmt_client_setup_async(mgmt_client, *, node, timeout):
+        """Open the management link without dispatching a request.
+
+        :param AMQPClient mgmt_client: Client used to open the management link.
+        :keyword bytes node: Management target.
+        :keyword int timeout: Timeout in seconds.
         """
 
     @staticmethod